// job-portal-fe-admin/src/model/userDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserDto {
  email?: string;
  id?: string;
  nui?: string;
  lastName?: string;
  name?: string;
  scope?: UserDto.ScopeEnum;
}

export namespace UserDto {
  export type ScopeEnum = 'ADMIN' | 'COMPANY' | 'REGULAR_USER';
  export const ScopeEnum = {
    ADMIN: 'ADMIN' as ScopeEnum,
    COMPANY: 'COMPANY' as ScopeEnum,
    REGULARUSER: 'REGULAR_USER' as ScopeEnum
  };
}